separate basic information into separate component
